Removing unnecessary files, changing the project title, and creating the 'Details' component
Cleaning the project by removing the files that won't be used, modifying the 'index.html' file by removing unnecessary code and changing the project title to 'Rocket Notes,' and removing unnecessary lines of code from the 'main.jsx' file. Finally, creating the 'Pages' folder and modifying the 'App.jsx' file by removing the default code, moving the file to the 'Pages' folder, renaming it to 'Details.jsx,' and modifying the rendered content.