Installing the 'styled components' library and applying styles through JavaScript
Installing the 'styled components' library to work with component styling using JavaScript. Transforming the 'styles.css' file of the 'Details' component into the 'styles.js' file using the CSS-in-JS technique and the syntax of 'styled components' to apply styles and create styled components with JavaScript code. Finally, updating the 'index.jsx' file of the 'Details' component to correctly import the styles and styled components.