Creating the 'styles' folder and the 'global.js' and 'theme.js' files
Creating the 'styles' folder at the root of the project to store the JavaScript files containing the global styles of the application. Creating the 'theme.js' file, exporting the object representing the application's theme with the colors that will be used. Creating the 'global.js' file, which, through the 'styled components' createGlobalStyle component, exports the general styles of the application, applying the background color and default text color dynamically. Finally, updating the 'main.js' file to import the components and apply the styles to the application