Creating the 'Button' component and applying it to the 'Details' page
Creating the 'components' folder and the 'Button' component, along with the 'index.jsx' file for this component and the 'styles.js' file. Inside the 'styles.js' file, creating the styled 'Container' component, containing the button styles. Finally, applying the created component to the 'Details' page and passing properties to this component.