'Details' page evolving by implementing the 'Header' component.
Continuing the development of the details page, creating the 'Header' component, representing the page header, with its respective 'index.jsx' and 'style.js' files. Within the 'style.js' file of the 'Header' component, creating the sub-components 'Container' and 'Profile,' representing the header as a whole and the user information and photo part, respectively. Finally, applying the 'Header' component to the 'Details' page component and implementing the "Grid layout" technique in the component.