Installing the 'react-icons' library and creating and implementing the 'Section' and 'Tag' components
Installing the 'react-icons' library to work with icons in the application, using a 'Logout' icon in the 'Header' component. Creating the 'Section' component to represent a section on the 'Details' page and creating the 'Tag' component to represent a content marker for a note. Finally, applying the styles of these components through their respective 'style.js' files and implementing them in the 'Details' page component