Creating the 'ButtonText' component and finalizing the 'Details' page
Creating the last component of the 'Details' page, the 'ButtonText' representing a button without borders and a background color. Completing the 'Details' page by wrapping the main content of the page in a 'main' tag and encapsulating the content within the stylized 'Content' component. Finally, adding a title and a paragraph to the page, and applying the last styles of the component