Creating the 'Note' component and finalizing the 'Home' page.
Creating the 'Note' component that represents a note created by the user and receives an object containing the user-provided note data, such as the note's title and its list of tags. Applying styles to the 'Note' component and implementing it on the 'Home' page. Finally, completing the initial page component by creating a new section with the notes created by the user