Building the registration page ('SignUp') and beginning the development of the user profile page
Creating the registration page using the 'SignUp' component, reusing the structure and styles from the 'SignIn' page and making necessary content changes. Initiating the development of the user profile page with the 'Profile' component. Finally, building the header of the 'Profile' page, starting the construction of the form by adding the input fields and the submit button and applying the initial styles to the header and the form.