Finishing the 'Profile' page, starting the creation of the 'New' page, and creating the 'TextArea' component.
Completing the development of the 'Profile' page, creating the stylized 'Avatar' sub-component, which includes the user's photo and the option to change the profile picture through a 'file' input, and applying the final styles to the page. Initiating the development of the 'New' page component for users to record a new note, adding the page header, and beginning the construction of the form. Lastly, adding the initial form fields and creating the 'TextArea' component, representing a multi-line text field for the user to input additional notes about the note to be created