Creating the routes and implementing navigation between the pages of the application
Creating the application routes within the 'routes' page in the project's source code and establishing navigation between the project's pages using the 'react-router-dom' library and its components. Creating the 'routes' folder with the 'app.routes.jsx' file, which contains the main application routes ('Home', 'New', 'Profile', and 'Details'), and the 'auth.routes.jsx' file responsible for managing the authentication routes of the application ('SignIn' and 'SignUp'). Implementing navigation between the application's pages using the 'Link' component from 'react-router-dom' and modifying the pages and components to make use of the 'Link' component. Finally, consolidating the route components into a single component called 'Routes' in the 'index.jsx' file within the 'routes' folder, and applying this component in the 'main.jsx' file of the application