Starting the integration with the application's Back-End, creating the registration logic
Starting the Front-End and Back-End integration phase. Installing the 'axios' library for making HTTP requests to fetch and send data to the API, creating the 'services' folder, and the 'api.js' file containing the configurations that will be passed to axios, including the API's address. Developing the user registration logic in the 'SignUp' component, capturing the data entered by the user in the input fields and storing them in states, and upon submitting the form, the stored data is sent to the API via an HTTP request. Lastly, creating the example file 'myContext.js' and using contexts in the application