Implementing route redirection and persisting user data sent by the API through localStorage
Continuing the development of the project, implementing the redirection of application routes according to user data. Validating whether the user has performed authentication and whether the user data was successfully obtained from the API. If so, the 'Routes' component will display the application routes ('AppRoutes'), which are the main routes of the application. Otherwise, the 'Routes' component will display the authentication routes ('AuthRoutes'), allowing the user to complete the registration and/or login process and gain access to the application. Finally, persisting the user data sent by the API during authentication in the user's browser in a permanent manner using localStorage, and using the 'useEffect' hook to retrieve the stored data and store it in a state for use throughout the application via the authentication context.