Implementing the logout logic and loading user data into the 'Profile' component.
Updating the authentication context in the 'auth.jsx' file within the 'hooks' folder, creating the 'signOut' function that contains the logout logic, allowing the user to log out of the application by clearing the data saved in localStorage. Updating the 'Header' component, implementing the 'signOut' function through the 'Logout' component. Finally, loading the user's data on the 'Profile' page using the 'Input' components, to later implement the functionality of updating the user's data