Creating the logic to update user data and starting the necessary code to create a new note.
Building the necessary code to update user data in the database through the 'Profile' page. Constructing the 'updateProfile' function in the authentication hook, containing the logic to update user data, and making it available to other components through the authentication context. Updating the 'Profile' component, capturing the data entered by the user in the input fields and using the 'updateProfile' function to update the user's information with the new data. Also, creating the logic to update the user's avatar using an image file provided by them and setting a default avatar for all users. Finally, starting to build the logic for recording a new note, assembling the code responsible for allowing the user to add and remove links related to a note