Implementing the functionality to remove a note, adjusting the application's route navigation, and making adjustments to the 'Profile' component
Developing the final codes of the application. Implementing the functionality to delete a note created by the user through the 'Details' component, removing the note data from the database and redirecting the user to the home page. Adjusting the navigation between the application routes, returning to the previous page when necessary, instead of sending the user directly to the home page. Fine-tuning the logic of the 'Profile' component, correctly saving the user's avatar path, merging the existing data with the new data provided by the user. Updating the 'updateProfile' method of the authentication hook to avoid saving sensitive data in local storage, such as the user's password. Finally, testing the application by registering a new user and verifying all functionalities and validations